+++ zcash_primitives/CHANGELOG.md
@@ -7,7 +7,8 @@ and this library adheres to Rust's notion of
 
 ## [Unreleased]
 ### Added
-- `zcash_primitives::legacy::keys::NonHardenedChildIndex`
+- `zcash_primitives::legacy::keys::{NonHardenedChildIndex, TransparentKeyScope}`
+- `zcash_primitives::legacy::keys::AccountPrivKey::derive_secret_key`
 - Dependency on `bellman 0.14`.
 - `zcash_primitives::consensus::sapling_zip212_enforcement`
 - `zcash_primitives::transaction`:
@@ -127,7 +128,8 @@ and this library adheres to Rust's notion of
 - `zcash_client_backend` changes related to `local-consensus` feature:
   - added tests that verify `zip321` supports Payment URIs with `Local(P)`
   network parameters.
-- `zcash_primitives::legacy::keys::derive_external_secret_key` parameter type changed from `u32` to `NonHardenedChildIndex`.
+- `zcash_primitives::legacy::keys::{derive_external_secret_key, derive_internal_secret_key}` 
+  arguments changed from `u32` to `NonHardenedChildIndex`.
 
 ### Removed
 - `zcash_primitives::constants`:
